+++ src/apiHelper.js
@@ -24,7 +24,7 @@ const getQueryString = (config, existingQs = {}) => {
     uuid: config.deviceId
   };
   if (config.gid) { defaultQs.gid = config.gid; }
-  const qsObject = { ...existingQs, ...defaultQs };
+  const qsObject = Object.assign({}, existingQs, defaultQs);
   const queryString = qs.stringify(qsObject);
   return queryString;
 };
@@ -38,11 +38,11 @@ const getRequestUrlWithQueryString = (path, config) => {
 };
 
 const getExtraHeaders = (options) => {
-  return { ...getForwardedForHeader(options), ...getSessionHeader(options) };
+  return Object.assign({}, getForwardedForHeader(options), getSessionHeader(options));
 };
 
 const getFetch = (path, config) => {
-  const headers = { ...defaultHeaders, ...getExtraHeaders(config) };
+  const headers = Object.assign({}, defaultHeaders, getExtraHeaders(config));
   const requestUrl = getRequestUrlWithQueryString(path, config);
   config.log(`Sending a GET request to: ${requestUrl} with the following headers`, headers);
   return fetch(requestUrl, { credentials, headers })
@@ -73,11 +73,11 @@ export const grabRaw = (path, config) => {
 };
 
 export const post = (path, config, body = {}) => {
-  const headers = {
-    ...defaultHeaders,
-    ...getContentTypeHeader(),
-    ...getExtraHeaders(config)
-  };
+  const headers = Object.assign({},
+    defaultHeaders,
+    getContentTypeHeader(),
+    getExtraHeaders(config)
+  );
   const requestUrl = getRequestUrlWithQueryString(path, config);
   config.log(`Sending a POST request to: ${requestUrl}. With the following headers and body: `, headers, body);
   const options = {